refactor package structure
Move feature extractors and auxiliary functions in util.py .
Add demo function.
